+++ Orbitersdk/stackEditor/StackExport.cpp
@@ -38,7 +38,7 @@ void StackExport::prepareExportData()
 				int dockedvidx = _data->stack->getIndexOfVessel(Helpers::getVesselByUID(v->dockingPorts[j].dockedTo.vesselUID));
 				//connections are only remembered down the queue. 
 				//Vessels appearing earlier in the queue must not know that they're docked to vessels higher up
-				if (dockedvidx < i)
+				if (dockedvidx < (int)i)
 				{ 
 					DockPortExport newd;
 					newd.dockedToVessel = dockedvidx;
